requirements.txt
apache-airflow==2.2.0
pandas
boto3
apache-airflow-providers-postgres